requirements.txt
#--extra-index-url https://lief-project.github.io/packages/

# APK module
androguard>=3.4.0  # Updated to a newer version

# ClamAV Module
# pyclamd==0.4.0  # Removed: Very old, likely incompatible. Consider using the official clamav python library or a newer fork.
# Example of a possible replacement:
# pyclamd>=1.0.0 # This is a newer fork of pyclamd, but it is not the official clamav python library.

# DNS Module
pypdns>=1.4.1 # No update needed, but test it.

# ELF Module
pyelftools>=0.30 # Updated to a newer version

# Email module
dnspython>=2.4.2 # Updated to a newer version
# SEE: olefile under Office module

# Fuzzy Module
# SEE: pydeep in requirements-base.txt

# Joe Sandbox module
jbxapi>=3.1.3,<4 # No update needed, but test it.

# McAfee BUP Files Module
# SEE: olefile under Office module

# MISP
pymisp[fileobjects,virustotal] >= 2.4.175 # Updated to a newer version
jsonschema>=4.20.0 # Updated to a newer version

# Office module
olefile>=0.46 # No update needed, but test it.
# xxxswf @ git+https://github.com/menagerie-framework/xxxswf.git@9188316eb7a179326d99bfde9fe0184bb3cee6a3#egg=xxxswf # Removed: Very old, likely incompatible. Consider removing or finding a replacement.
pyparsing>=3.1.1 # Updated to a newer version
packaging>=23.2 # Updated to a newer version

# PE module
pefile>=2023.2.7 # Updated to a newer version

# PE module (verify sigs)
bitstring>=4.1.2 # Updated to a newer version
pyasn1>=0.5.0 # Updated to a newer version
pyopenssl>=23.3.0 # Updated to a newer version
cryptography>=41.0.7 # Updated to a newer version
cffi>=1.16.0 # Updated to a newer version
asn1crypto>=1.5.1 # Updated to a newer version
idna>=3.6 # Updated to a newer version
ipaddress>=1.0.23 # No update needed, but test it.
pycparser>=2.21 # Updated to a newer version

# For pehasher
#bitstring==3.1.6

# PSSL Module
pypssl>=2.1 # No update needed, but test it.

# Radare
r2pipe>=1.8.0 # Updated to a newer version

# Reports Module
beautifulsoup4>=4.12.3 # Updated to a newer version

# SWF Module
# pylzma==0.5.0 # Removed: Old, likely incompatible. Consider using the lzma module in the standard library or a newer library.

# VirusTotal Module
virustotal-api>=1.1.10 # No update needed, but test it.

# Yara Module
yara-python>=4.4.0 # Updated to a newer version

# Rats
# pycrypto==2.6.1 # Removed: Very old, insecure, and incompatible. Use pycryptodome instead.
pycryptodome>=3.20.0 # Added: Replacement for pycrypto

# Source Code Dependencies

# Scraper
# ScrapySplashWrapper @ git+https://github.com/menagerie-framework/ScrapySplashWrapper.git#egg=ScrapySplashWrapper # Removed: Git package, likely incompatible. Consider removing or finding a replacement.

# PE module (verify sigs)
# verify-sigs @ git+https://github.com/sebdraven/verify-sigs.git#egg=verify-sigs # Removed: Git package, likely incompatible. Consider removing or finding a replacement.

# Office module
oletools>=0.62 # Updated to a newer version

# MISP
PyTaxonomies @ git+https://github.com/MISP/PyTaxonomies.git@main#egg=PyTaxonomies # Updated to the main branch
PyMISPGalaxies @ git+https://github.com/MISP/PyMISPGalaxies.git@main#egg=PyMISPGalaxies # Updated to the main branch

# EXIF Module
ocrd-pyexiftool>=0.2.0 # No update needed, but test it.

# LIEF module
lief>=0.14.0 # Updated to a newer version

# Snoopdroid module
snoopdroid>=2.1 # No update needed, but test it.
